feat: :lipstick: rewrite index layout with grid
The grid is divided into 9 cells.
The columns have 10%, 80%, 10% width respectively.
The top and bottom rows fit their content, while the middle takes the rest of the space.

The header and footer occupy their whole row by spanning across all columns, so for the pages to fully occupy the middle they also get the col-span-full class